is855/thumbnails (#950)
Now we can include URLs in the thumbnail field of a node.
Thumbnails removed from all kinds of plots. Rawgraph is generating problematic markup.
Added a 'give us feedback function'. Open issue in GitHub is back in the user menu.
Node icons sitting in one line and aligned to the left (requested by Esra and Nik)
Changed FilePicker's bottom bar look.
Changed loading icon to be the same we are using in the loading nodes in the workbench (this was kind of tricky).