[bugfix] Delete studies (#1226)
Now using Promise.all to wait for every delete call to resolve before fetching the studies again.
This didn't solve the problem, which was caused by the delete endpoint, not waiting for the delete process to finish before sending the response. This caused the front end to try to fetch the studies before they were actually deleted.
Now the backend stops the services and removes the study's data in parallel but waits for the studies to be deleted from the database before sending a response.