Adds a new receipe to upgrade all dev tools
Once more, I had the wrong version of qx compiler. Needed a one-key-solutino
to get all tooling right in place. Avoided adding it to ``make devenv`` since
it might slow it down and this target is heavily used. For that reason, I
thought it might be better to add an extension to it. In a similar fashion to
clean, clean-images and clean-all