Is1273/groups api (#1276)
- new catalog service
- adds new table of dag to simcore-postgres-database package
- webserver: new catalog-gateway subsystem
- api extends to

Maintenance:
- web API doc: documented and created tags in openapi
- makefile: webserver and catalog service makefile have all build and test target flavours
- adds shellcheck tooling in scripts/shellcheck and vscode configuration in .vscode-template/settings.json
- NEW standard makefile in simcore-postgres-database
- highlight: has a setup to create migration commits
- IMPORTANT: this PR requires upgrading deployed databases


* Builds catalog image
* Adds a new receipe to upgrade all dev tools
Once more, I had the wrong version of qx compiler. Needed a one-key-solutino
to get all tooling right in place. Avoided adding it to ``make devenv`` since
it might slow it down and this target is heavily used. For that reason, I
thought it might be better to add an extension to it. In a similar fashion to
clean, clean-images and clean-all
* Enhanced build targets
By creating multiline variables we can deal with many variants more reliably
Running qx compiler every time was too time consuming and some of the variants (e.g. no-cache) etc were missing in devel