e2e upgrade (#2352)
Uses a combination of puppeteer, jest-puppeteer and jest that appears to be more stable. Tests seem to run a bit more consistent across platforms (MacOS, Linux, etc) with more consistent results. And screenshots don't hang.
Also checks for file names instead of only number of files.
Adds a different, not GUI method to delete the studies.
When puppeteer waits, it is possible to include a reason that will be logged.
Flash messages are logged